requirements.txt
Authlib==1.2.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
decorator==5.1.1
distlib==0.3.6
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.9.0
Flask==2.2.3
Flask-Authlib-Client==0.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
oauthlib==3.2.2
openpyxl==3.1.2
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.0.0
ply==3.11
prompt-toolkit==3.0.36
protobuf==4.22.1
psycopg2-binary~=2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.2
stone==3.3.1
traitlets==5.9.0
tzdata==2022.7
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
